refactor: standardize file and folder naming
- Rename TSX files to PascalCase
- Rename folders to kebab-case
